backend/requirements.txt
fastapi
uvicorn
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
requests
python-multipart
itsdangerous
